Moved modules reg to /src level
