--- src/themes/theme-starter/modules/index.ts
@@ -1,29 +1,0 @@
-import { VueStorefrontModule, VueStorefrontModuleConfig } from '@vue-storefront/core/modules'
-import { Cart } from '@vue-storefront/core/modules/cart'
-import { Review } from '@vue-storefront/core/modules/review'
-import { Mailer } from '@vue-storefront/core/modules/mailer'
-import { Mailchimp } from '@vue-storefront/core/modules/mailchimp'
-import { Example } from '@vue-storefront/core/modules/module-template'
-import { Notification } from '@vue-storefront/core/modules/notification'
-import { DroppointShipping } from '@vue-storefront/core/modules/droppoint-shipping'
-
-// Some modules  that still needs API refactoring are  temporary registered in core
-// This is how you can adjust any module with application-specific behavior
-const extendedExample: VueStorefrontModuleConfig = {
-  key: 'extend',
-  afterRegistration: function(Vue, config) {
-    console.info('Hello, im extended now!')
-  }
-}
-
-Example.extend(extendedExample)
-
-export const registerModules: VueStorefrontModule[] = [
-  Cart,
-  Review,
-  Mailer,
-  Mailchimp,
-  Notification,
-  DroppointShipping,
-  Example
-]
